модуль product  test_product
