+++ htmlcov/index.html
@@ -11,7 +11,7 @@
 <header>
     <div class="content">
         <h1>Coverage report:
-            <span class="pc_cov">92%</span>
+            <span class="pc_cov">79%</span>
         </h1>
         <aside id="help_panel_wrapper">
             <input id="help_panel_state" type="checkbox">
@@ -54,7 +54,7 @@
         </h2>
         <p class="text">
             <a class="nav" href="https://coverage.readthedocs.io/en/7.9.2">coverage.py v7.9.2</a>,
-            created at 2025-07-09 11:53 +0300
+            created at 2025-07-09 12:36 +0300
         </p>
     </div>
 </header>
@@ -78,11 +78,39 @@
                 <td class="right" data-ratio="0 0">100%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html">src\main.py</a></td>
-                <td>138</td>
-                <td>4</td>
+                <td class="name left"><a href="z_145eef247bfb46b6_base_category_order_py.html">src\base_category_order.py</a></td>
+                <td>5</td>
+                <td>1</td>
                 <td>0</td>
-                <td class="right" data-ratio="134 138">97%</td>
+                <td class="right" data-ratio="4 5">80%</td>
+            </tr>
+            <tr class="region">
+                <td class="name left"><a href="z_145eef247bfb46b6_base_product_py.html">src\base_product.py</a></td>
+                <td>12</td>
+                <td>3</td>
+                <td>0</td>
+                <td class="right" data-ratio="9 12">75%</td>
+            </tr>
+            <tr class="region">
+                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html">src\main.py</a></td>
+                <td>105</td>
+                <td>29</td>
+                <td>0</td>
+                <td class="right" data-ratio="76 105">72%</td>
+            </tr>
+            <tr class="region">
+                <td class="name left"><a href="z_145eef247bfb46b6_print_mixin_py.html">src\print_mixin.py</a></td>
+                <td>5</td>
+                <td>0</td>
+                <td>0</td>
+                <td class="right" data-ratio="5 5">100%</td>
+            </tr>
+            <tr class="region">
+                <td class="name left"><a href="z_145eef247bfb46b6_product_py.html">src\product.py</a></td>
+                <td>47</td>
+                <td>0</td>
+                <td>0</td>
+                <td class="right" data-ratio="47 47">100%</td>
             </tr>
             <tr class="region">
                 <td class="name left"><a href="z_145eef247bfb46b6_utils_py.html">src\utils.py</a></td>
@@ -95,10 +123,10 @@
         <tfoot>
             <tr class="total">
                 <td class="name left">Total</td>
-                <td>156</td>
-                <td>12</td>
+                <td>192</td>
+                <td>41</td>
                 <td>0</td>
-                <td class="right" data-ratio="144 156">92%</td>
+                <td class="right" data-ratio="151 192">79%</td>
             </tr>
         </tfoot>
     </table>
@@ -110,7 +138,7 @@
     <div class="content">
         <p>
             <a class="nav" href="https://coverage.readthedocs.io/en/7.9.2">coverage.py v7.9.2</a>,
-            created at 2025-07-09 11:53 +0300
+            created at 2025-07-09 12:36 +0300
         </p>
     </div>
     <aside class="hidden">
